Extracting linelist methods into helper so other roles can access, and making role agnostic when retrieving user patients. Updating Move to Household language and updating formatting of name to be consistent with dashboard pages (link + hoh badge).
Removing unnecessary indexes added through migration and cleaning up.

Fixing displayed name in modal.
