+++ app/javascript/packs/stylesheets/custom_table.scss
@@ -28,6 +28,3 @@
     background-color: white;
 }
 
-.no-data-label {
-    min-height: 300px !important;
-}
