--- db/migrate/20210108232031_add_chain_indexes_to_patients.rb
@@ -1,6 +1,0 @@
-class AddChainIndexesToPatients < ActiveRecord::Migration[6.0]
-  def change
-    add_index :patients, [:purged, :isolation, :last_name, :first_name, :jurisdiction_id], name: 'patients_index_chain_five_1'
-    add_index :patients, [:purged, :isolation, :jurisdiction_id], name: 'patients_index_chain_three_2'
-  end
-end
